+++ changelog.md
@@ -13,6 +13,8 @@ Codion Change Log
 - AbstractEntityConnectionProvider.validateConnection() replaced with validConnection(), connection() no longer final.
 ### is.codion.framework.db.local
 - LocalEntityConnectionProvider now overrides connection() with LocalEntityConnection return type.
+### is.codion.framework.model
+- DefaultEntityTableConditionModel bug fixed, inCondition() and notInCondition() now return isNull() and isNotNull() respectively when no operands are specified.
 ### is.codion.swing.framework.model
 - SwingEntityTableModel.isCellEditable() now final, protected editable() added.
 ### is.codion.swing.framework.ui